[ubitx] port over the "don't spit out harmonics during TX" patch from the bitx40 firmware fork.
This quietens the main VFO oscillator before enabling TX and then
sleeps a little bit to give the circuit time to settle before flipping
over to TX.  This prevents a spat of harmonics going out when you key
up CW.

Do the same on RX too - relays are slow, and this microprocessor runs
at a few MHz.  Give the circuit time to settle before continuing.
Calls to setFrequency() re-enable the VFO.